refactor: move selector module
