+++ avilla/core/trait/context.py
@@ -8,7 +8,7 @@ from typing import Any, Awaitable, Callable, MutableMapping, TypeVar, overload
 from typing_extensions import Concatenate, ParamSpec, TypeAlias, Unpack
 
 from ..context import Context
-from ..utilles.selector import Selector
+from ..selector import Selector
 from ..metadata import Metadata, MetadataBound, MetadataOf, MetadataRoute
 from ..resource import Resource
 from . import (
