+++ avilla/elizabeth/impl/friend.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from avilla.core.trait.context import bounds, implement
-from avilla.core.utilles.selector import Selector
+from avilla.core.selector import Selector
 from avilla.spec.core.message import MessageRevoke, MessageSend
 
 if TYPE_CHECKING:
