+++ avilla/elizabeth/impl/group.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING
 from avilla.core.message import Message
 from avilla.core.trait.context import bounds, implement, pull
 from avilla.core.exceptions import permission_error_message
-from avilla.core.utilles.selector import Selector
+from avilla.core.selector import Selector
 from avilla.spec.core.message import MessageRevoke, MessageSend
 from avilla.spec.core.privilege import MuteAllTrait, MuteTrait, Privilege
 from avilla.spec.core.profile import Nick, Summary, SummaryTrait
