+++ avilla/spec/core/message/event.py
@@ -7,7 +7,7 @@ from graia.amnesia.message import __message_chain_class__
 
 from avilla.core.event import AvillaEvent
 from avilla.core.message import Message
-from avilla.core.utilles.selector import Selector
+from avilla.core.selector import Selector
 
 if TYPE_CHECKING:
     from graia.broadcast.interfaces.dispatcher import DispatcherInterface
