+++ avilla/spec/core/message/skeleton.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from graia.amnesia.message import __message_chain_class__
 
 from avilla.core.trait import Fn, Trait
-from avilla.core.utilles.selector import Selector
+from avilla.core.selector import Selector
 
 # MessageFetch => rs.pull(Message, target=...)
