+++ avilla/spec/core/privilege/skeleton.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from datetime import timedelta
 
 from avilla.core.trait import Fn, Trait
-from avilla.core.utilles.selector import Selector
+from avilla.core.selector import Selector
 
 
 class PrivilegeTrait(Trait):
