+++ avilla/spec/core/resource/event.py
@@ -8,7 +8,7 @@ from typing_extensions import get_origin
 
 from avilla.core.event import AvillaEvent
 from avilla.core.resource import BlobResource, Resource
-from avilla.core.utilles.selector import Selector
+from avilla.core.selector import Selector
 
 if TYPE_CHECKING:
     from graia.broadcast.interfaces.dispatcher import DispatcherInterface
